+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - added guide for migrating from tmuxinator
 - fixed a bug for restoring commands on tmux 2.5 (and probably tmux 2.4)
 - do not create another resurrect file if there are no changes (credit @vburdo)
+- allow using '$HOSTNAME' in @resurrect-dir
 
 ### v2.4.0, 2015-02-23
 - add "tmux-test"